requirements.txt
python-telegram-bot==21.10
openai==1.65.2
requests==2.32.3
